requirements.txt
arrow==0.14.4
astroid==2.2.5
blessed==1.15.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cloudinary==1.11.0
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.0.4
django-anymail==6.1.0
django-cors-headers==3.0.2
django-picklefield==2.0
django-q==1.0.1
django-templated-mail==1.1.1
django-webpack-loader==0.6.0
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mock==3.0.5
mysqlclient==1.4.6
pip-licenses==1.15.2
postgres==2.2.2
psycopg2-binary==2.8.3
PTable==0.9.2
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
sparkpost==1.3.6
sqlparse==0.3.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.3
wcwidth==0.1.7
whitenoise==5.0.1
wrapt==1.11.2
